In progress

+++ masterkube/guestinfos/cloud-init-clean.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-
-rm /etc/cloud/cloud.cfg.d/50-curtin-networking.cfg
+[ -f /etc/cloud/cloud.cfg.d/50-curtin-networking.cfg ] && rm /etc/cloud/cloud.cfg.d/50-curtin-networking.cfg
+rm /etc/netplan/*
 cloud-init clean
 rm /var/log/cloud-ini*
-shutdown now
+shutdown -P now
